attendance marking

+++ database/migrations/2019_01_08_174904_create_students_table.php
@@ -19,7 +19,7 @@ class CreateStudentsTable extends Migration
             $table->string('name');
             $table->text('email');
             $table->string('level');
-            $table->unsignedInteger('attendance');
+            $table->tinyInteger('attendance'); // 0 = absent , 1 = present , -1 = not marked
             $table->timestamps();
         });
     }
